+++ ServiceContract/Dtos/TemperatureDto.cs
@@ -6,11 +6,11 @@ namespace ServiceContract.Dtos
 {
     public class TemperatureDto
     {
-        public int Minimal { get; set; }
-        public int Maximal { get; set; }
-        public int Afternoon { get; set; }
-        public int Night { get; set; }
-        public int Evening { get; set; }
-        public int Morning { get; set; }
+        public double Minimal { get; set; }
+        public double Maximal { get; set; }
+        public double Afternoon { get; set; }
+        public double Night { get; set; }
+        public double Evening { get; set; }
+        public double Morning { get; set; }
     }
 }
